configurable duration + move config
