+++ src/main/java/at/hannibal2/skyhanni/config/features/combat/DeployableConfig.kt
@@ -5,12 +5,11 @@ import at.hannibal2.skyhanni.config.FeatureToggle
 import at.hannibal2.skyhanni.config.core.config.Position
 import at.hannibal2.skyhanni.features.combat.DeployableType
 import at.hannibal2.skyhanni.utils.ConfigUtils.jumpToEditor
-import at.hannibal2.skyhanni.utils.StringUtils.allLettersFirstUppercase
 import com.google.gson.annotations.Expose
+import io.github.notenoughupdates.moulconfig.annotations.Accordion
 import io.github.notenoughupdates.moulconfig.annotations.ConfigEditorBoolean
 import io.github.notenoughupdates.moulconfig.annotations.ConfigEditorButton
 import io.github.notenoughupdates.moulconfig.annotations.ConfigEditorDraggableList
-import io.github.notenoughupdates.moulconfig.annotations.ConfigEditorSlider
 import io.github.notenoughupdates.moulconfig.annotations.ConfigLink
 import io.github.notenoughupdates.moulconfig.annotations.ConfigOption
 
@@ -41,38 +40,9 @@ class DeployableConfig {
     )
 
     @Expose
-    @ConfigOption(name = "Missing Deployable Warning", desc = "Warn when a deployable is missing.")
-    @ConfigEditorBoolean
-    @FeatureToggle
-    var warnMissingDeployable: Boolean = false
-
-    @Expose
-    @ConfigOption(
-        name = "Deployable Warnings",
-        desc = "Warn when the required deployable is missing: Flux for Slayer, Umberella for Trophy Fishing, and Lantern for Mineshafts.",
-    )
-    @ConfigEditorDraggableList
-    val warningTypes: MutableList<WarningType> = mutableListOf(
-        WarningType.SLAYER,
-        WarningType.TROPHY_FISHING,
-        WarningType.MINESHAFT,
-    )
-
-    enum class WarningType(val displayName: String, val deployableType: DeployableType) {
-        SLAYER("Slayer Boss Fight", DeployableType.FLUX),
-        TROPHY_FISHING("Trophy Fishing", DeployableType.UMBERELLA),
-        MINESHAFT("Mineshaft", DeployableType.LANTERN),
-        ;
-
-        override fun toString(): String {
-            return displayName + " (${deployableType.name.allLettersFirstUppercase()})"
-        }
-    }
-
-    @Expose
-    @ConfigOption(name = "Warning Delay", desc = "Delay before warning to place deployable.")
-    @ConfigEditorSlider(minValue = 0f, maxValue = 10f, minStep = 1f)
-    var warningDelay: Int = 5
+    @ConfigOption(name = "Deployable Reminder", desc = "")
+    @Accordion
+    val deployableReminder = DeployableReminderConfig()
 
     @ConfigOption(
         name = "Flare Display",
@@ -91,8 +61,4 @@ class DeployableConfig {
     @Expose
     @ConfigLink(owner = DeployableConfig::class, field = "enabled")
     val position: Position = Position(-160, -70)
-
-    @Expose
-    @ConfigLink(owner = DeployableConfig::class, field = "warnMissingDeployable")
-    val warningPosition: Position = Position(10, 10)
 }
